Add functions for basic configuration verification
Added functions in config.py to verify that the config file has the
necessary structure and that each section has all the necessary values.
At the moment this not a very advanced check but provides a decent
structure to build on if more is required in the future. A basic
true/false check is run everytime the program is started unless the
'--skip-config-check' flag is used. If there is a problem, the program
will quit. Skipping the config check may cuase unhandled exceptions to
be thrown becuase other methods will assume that config has all the
values it needs. To run a standalone verification of the config, use the
'--check-config' flag. This will print every error it finds in the
configuration file before exiting.