Add the 'skip_quick' option to settings
Moved the setting to define a quick test group to 'quick_group'. At
least in my uses, the quick group is duplicates from other groups that
don't need to run twice if I'm testing every host. So when this option
is enabled, a normal run of netrep will skip over the quick test group.
If this is true and the '-q' flag is given, only the quick test group
will be run.